+++ CHANGES.md
@@ -216,3 +216,4 @@
 - Extra Menu: text adjustment
 - Pre-Release Checker: change filename with latest commit date and ID
 - `StarkeyOtherBrands.bat`
+- Offline documents